assignemt 2/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
duckduckgo-search>=6.3.0
ddgs>=9.0.0
wikipedia>=1.4.0
python-dotenv>=1.0.1
streamlit>=1.37.0
reportlab>=4.2.0
